finish converting widget
