+++ src/lib/zodSchemas.ts
@@ -289,3 +289,38 @@ export const widgetSchemaBase = z.object({
   isPublished: z.boolean().optional(),
 });
 export const widgetSchemaBasePartial = widgetSchemaBase.partial();
+
+// Split out the refinement function so we can reuse it to compose a partial schema
+// below
+function refinementFunc(
+  val: z.infer<typeof widgetSchemaBase> | z.infer<typeof widgetSchemaBasePartial>,
+  ctx: z.RefinementCtx
+) {
+  if (val.isTemplate) {
+    if (!val.description) {
+      ctx.addIssue({
+        code: z.ZodIssueCode.custom,
+        message: `Description is required`,
+        path: ['description'],
+      });
+    }
+  }
+}
+export const widgetSchema = widgetSchemaBase.superRefine(refinementFunc);
+export type WidgetSchema = typeof widgetSchema;
+
+export const widgetSchemaPartial = widgetSchemaBasePartial.superRefine(refinementFunc);
+export type WidgetSchemaPartial = typeof widgetSchemaPartial;
+
+export const widgetTemplateSchema = z.object({
+  name: z.string().min(1, { message: 'Name is required' }),
+  description: z.string().min(1, { message: 'Description is required' }),
+});
+export type WidgetTemplateSchema = typeof widgetTemplateSchema;
+
+export const datasetSchema = z.object({
+  type: z.enum(['line', 'bar']).default('line'),
+  color: z.string(),
+  name: z.string().min(1, { message: 'Name is required' }),
+});
+export type DatasetSchema = typeof datasetSchema;
