+++ src/routes/widget/+page.server.ts
@@ -2,7 +2,8 @@ import { getSignedUrlPromises } from '$lib/aws/s3';
 import { CustomQueryRepo, type CustomQueryResults } from '$lib/customQuery';
 import { prisma } from '$lib/prisma';
 import { getSessionOrRedirect } from '$lib/utils';
-import { WidgetRepo, widgetSchema } from '$lib/widget';
+import { WidgetRepo } from '$lib/widget';
+import { widgetSchema } from '$lib/zodSchemas';
 import { fail, redirect } from '@sveltejs/kit';
 import { superValidate } from 'sveltekit-superforms/server';
 import type { Actions, PageServerLoad } from './$types';
