+++ src/routes/widget/[id]/+page.server.ts
@@ -4,7 +4,8 @@ import { APIError } from '$lib/errors';
 import { prisma } from '$lib/prisma';
 import { TrainingCycleRepo } from '$lib/trainingCycle';
 import { getSessionOrRedirect } from '$lib/utils';
-import { WidgetRepo, widgetSchema, widgetTemplateSchema } from '$lib/widget';
+import { WidgetRepo } from '$lib/widget';
+import { widgetSchema, widgetTemplateSchema } from '$lib/zodSchemas';
 import { fail } from '@sveltejs/kit';
 import { superValidate } from 'sveltekit-superforms/server';
 import type { Actions, PageServerLoad } from './$types';
