+++ src/routes/widget/[id]/dataset/[datasetId]/+page.server.ts
@@ -1,6 +1,7 @@
 import { prisma } from '$lib/prisma';
 import { getSessionOrRedirect } from '$lib/utils';
-import { datasetSchema, WidgetRepo } from '$lib/widget';
+import { WidgetRepo } from '$lib/widget';
+import { datasetSchema } from '$lib/zodSchemas';
 import { fail } from '@sveltejs/kit';
 import { superValidate } from 'sveltekit-superforms/server';
 import type { Actions } from './$types';
